aaaaaa

+++ firmware/PlainFFT.cpp
@@ -94,7 +94,7 @@ void PlainFFT::windowing(double *vData, uint16_t samples) {
 	}
 }
 
-double PlainFFT::majorPeak(double *vD, uint16_t samples, double samplingFrequency) {
+uint16_t PlainFFT::majorPeak(double *vD, uint16_t samples, double samplingFrequency) {
 	double maxY = 0;
 	uint16_t IndexOfMaxY = 0;
 	for (uint16_t i = 1; i < ((samples >> 1) - 1); i++) {
